implemented stream encryption
